// reports/fast-input-decision-v2/baseline-before-frozen-production/src/domain/decision/decision.types.ts
export const DecisionAction = {
  FOLD: 'FOLD',
  CHECK: 'CHECK',
  CALL: 'CALL',
  BET: 'BET',
  RAISE: 'RAISE',
  ALL_IN: 'ALL_IN',
} as const;
export type DecisionAction = (typeof DecisionAction)[keyof typeof DecisionAction];